Implement boost strength method and calls; seems to work but need add strength bars/labels before testing further

+++ SumoBasho/Hero.h
@@ -10,7 +10,7 @@
 
 @interface Hero : SKSpriteNode
 
-@property int strength;
+@property CGFloat strength;
 
 + (id)heroWithImage:(NSString *)imageString name:(NSString *)name position:(CGPoint)position;
